+++ holoviews/__init__.py
@@ -29,7 +29,7 @@ from .element import *                                   # noqa (API import)
 from .element import __all__ as elements_list
 from .util import (extension, renderer, output, opts,    # noqa (API import)
                    render, save)
-from .util.ops import dim                                # noqa (API import)
+from .util.transform import dim                          # noqa (API import)
 
 # Suppress warnings generated by NumPy in matplotlib
 # Expected to be fixed in next matplotlib release
