+++ holoviews/plotting/bokeh/chart.py
@@ -12,7 +12,7 @@ from ...core import Dataset, OrderedDict
 from ...core.util import max_range, basestring, dimension_sanitizer, isfinite, range_pad
 from ...element import Bars
 from ...operation import interpolate_curve
-from ...util.ops import dim
+from ...util.transform import dim
 from ..util import compute_sizes, get_min_distance, dim_axis_label, get_axis_padding
 from .element import ElementPlot, ColorbarPlot, LegendPlot
 from .styles import (expand_batched_style, line_properties, fill_properties,
