+++ holoviews/plotting/bokeh/path.py
@@ -7,7 +7,7 @@ import numpy as np
 
 from ...core import util
 from ...element import Polygons
-from ...util.ops import dim
+from ...util.transform import dim
 from .element import ColorbarPlot, LegendPlot
 from .styles import (expand_batched_style, line_properties, fill_properties,
                      mpl_to_bokeh, validate)
