+++ holoviews/util/parser.py
@@ -17,7 +17,7 @@ import pyparsing as pp
 from ..core.options import Options, Cycle, Palette
 from ..core.util import merge_option_dicts
 from ..operation import Compositor
-from .ops import dim, norm
+from .transform import dim
 
 ascii_uppercase = 'ABCDEFGHIJKLMNOPQRSTUVWXYZ'
 allowed = '0123456789abcdefghijklmnopqrstuvwxyzABCDEFGHIJKLMNOPQRSTUVWXYZ!#$%&\()*+,-./:;<=>?@\\^_`{|}~'
@@ -36,8 +36,7 @@ class Parser(object):
     """
 
     # Static namespace set in __init__.py of the extension
-    namespace = {'np': np, 'Cycle': Cycle, 'Palette': Palette, 'dim': dim,
-                 'norm': norm}
+    namespace = {'np': np, 'Cycle': Cycle, 'Palette': Palette, 'dim': dim}
     # If True, raise SyntaxError on eval error otherwise warn
     abort_on_eval_failure = False
